requirements.txt
streamlit
pandas
gspread
google-auth
Pillow
pyzbar
opencv-python
pytz


